+++ qa_api/api_settings.py
@@ -1,0 +1,13 @@
+from flask import Flask, request
+from flask_caching import Cache
+
+config = {
+    "DEBUG": True,
+    "CACHE_TYPE": "SimpleCache",  # Flask-Caching related configs
+    "CACHE_DEFAULT_TIMEOUT": 600
+}
+
+app = Flask(__name__)
+# tell Flask to use the above defined config
+app.config.from_mapping(config)
+cache = Cache(app)
